refactor write results
